chore: remove eslint dependency
after agreed with Carlos Serrano. This project barely uses any javascript to lint and it makes it way more complex to maintain due to the need to handle breaking changes in both eslint and eslint-config-mailonline when they come. The simpler the better.

BREAKING CHANGE: remove eslint and eslint-confi-mailonline

Issues: MOL-29290